requirements.txt
anyio==3.6.2
certifi==2022.9.24
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
fastapi==0.86.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
pydantic==1.10.2
pymongo==4.3.2
python-dotenv==0.21.0
rfc3986==1.5.0
sniffio==1.3.0
starlette==0.20.4
tekore==4.5.0
typing_extensions==4.4.0
uvicorn==0.19.0
